Address PR feedback.  Undo move of some settings related code.  Only create semanticTokenLegend once.
