+++ Extension/src/LanguageServer/Providers/documentFormattingEditProvider.ts
@@ -15,7 +15,7 @@ export class DocumentFormattingEditProvider implements vscode.DocumentFormatting
 
     public async provideDocumentFormattingEdits(document: vscode.TextDocument, options: vscode.FormattingOptions, token: vscode.CancellationToken): Promise<vscode.TextEdit[]> {
         const settings: CppSettings = new CppSettings(vscode.workspace.getWorkspaceFolder(document.uri)?.uri);
-        if (settings.formattingEngine === "Disabled") {
+        if (settings.formattingEngine === "disabled") {
             return [];
         }
         await this.client.awaitUntilLanguageClientReady();
